fix uint/int bug, make roles get returned from view, & refactor

+++ pkg/utl/zlog/zlog.go
@@ -31,7 +31,7 @@ func (z *Log) Log(ctx echo.Context, source, msg string, err error, params map[st
 	params["source"] = source
 
 	if id, ok := ctx.Get("id").(uint); ok {
-		params["id"] = uint(id)
+		params["id"] = id
 		params["user"] = ctx.Get("username").(string)
 	}
